chore: [#7] Created separated folders for files from src
