+++ src/components/Navbar/Navbar.tsx
@@ -1,8 +1,8 @@
 import React, { useContext, useState } from 'react';
 
-import { CONTENT_TYPE, SHOW_PLACEHOLDER, MOVIE_PLACEHOLDER } from 'constantValues';
-import { MoviesShowsContext } from 'Context';
-import { NavbarState } from 'types';
+import { CONTENT_TYPE, SHOW_PLACEHOLDER, MOVIE_PLACEHOLDER } from 'constants/constantValues';
+import { MoviesShowsContext } from 'context/Context';
+import { NavbarState } from 'types/types';
 
 import './Navbar.css';
