--- src/helpers.tsx
@@ -1,1 +1,0 @@
-export const calculateAverageVote = (vote: number): number => Math.round(vote * 10) / 10;
