+++ src/helpers/calculateAverageVote.ts
@@ -1,0 +1,3 @@
+const calculateAverageVote = (vote: number): number => Math.round(vote * 10) / 10;
+
+export default calculateAverageVote;
