+++ src/pages/Home/Home.tsx
@@ -3,9 +3,9 @@ import React, { useContext } from 'react';
 import ItemCard from 'components/Item/ItemCard';
 import Loading from 'components/Loading/Loading';
 import Navbar from 'components/Navbar/Navbar';
-import { CONTENT_TYPE } from 'constantValues';
-import { MoviesShowsContext } from 'Context';
-import { IShow, IMovie, ItemType } from 'types';
+import { CONTENT_TYPE } from 'constants/constantValues';
+import { MoviesShowsContext } from 'context/Context';
+import { IShow, IMovie, ItemType } from 'types/types';
 
 import './Home.css';
